New blood request by doctor done

+++ src/main/java/blood_donation/domain/blood/Platelet.java
@@ -1,5 +1,6 @@
 package blood_donation.domain.blood;
 
+import javafx.beans.property.FloatProperty;
 import javafx.beans.property.IntegerProperty;
 
 import javax.persistence.DiscriminatorValue;
@@ -10,7 +11,7 @@ import java.time.LocalDate;
 @DiscriminatorValue(value = "platelet")
 public class Platelet extends Blood
 {
-    public Platelet(IntegerProperty quantity, BloodGroup bloodGroup, LocalDate lifeSpan)
+    public Platelet(FloatProperty quantity, BloodGroup bloodGroup, LocalDate lifeSpan)
     {
         super(quantity, bloodGroup, lifeSpan);
     }
